+++ libs/mb-utils/mb_square_root_vector.c
@@ -4,6 +4,6 @@
 void mb_square_root_vector(size_t length, float vect[length], float sqrt_vector[length]) {
     for (int i = 0; i < length; ++i) {
         float element = vect[i];
-        sqrt_vector[i] = sqrt(element);
+        sqrt_vector[i] = sqrtf(element);
     }
 }
